backend/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
APScheduler==3.11.0
arrow==1.3.0
attrs==24.3.0
awscli==1.37.6
beautifulsoup4==4.12.3
blinker==1.9.0
boto3==1.36.6
botocore==1.36.6
bs4==0.0.2
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.0
chromedriver-binary==131.0.6778.2.0
click==8.1.7
colorama==0.4.6
distro==1.9.0
docutils==0.16
dropbox==12.0.2
Flask==3.1.0
Flask-APScheduler==1.13.1
Flask-Cors==5.0.0
git-filter-repo==2.47.0
google-api-core==2.24.0
google-api-python-client==2.155.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
ics==0.7.2
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.2
jmespath==1.0.1
MarkupSafe==3.0.2
oauthlib==3.2.2
openai==1.57.4
outcome==1.3.0.post0
packaging==24.2
pluggy==1.5.0
ply==3.11
proto-plus==1.25.0
protobuf==5.29.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.3
pydantic_core==2.27.1
pyparsing==3.2.0
PySocks==1.7.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.7.2
s3transfer==0.11.2
selenium==4.27.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.37
stone==3.3.1
TatSu==5.12.2
tqdm==4.67.1
trio==0.27.0
trio-websocket==0.11.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.3
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
